budget_premium/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pandas>=2.2
openpyxl>=3.1
httpx>=0.27
python-multipart>=0.0.9
fpdf2>=2.7
matplotlib>=3.9
